HW-3 repository, remote data source interface 분리
