+++ wswon/AndroidArchitectureStudy/app/src/main/java/com/architecture/study/data/repository/CoinRepository.kt
@@ -1,10 +1,10 @@
-package com.architecture.study.data.source
+package com.architecture.study.data.repository
 
+import com.architecture.study.data.source.CoinRemoteDataSourceListener
 import com.architecture.study.network.model.MarketResponse
 import com.architecture.study.network.model.TickerResponse
 
-interface CoinDataSource {
-
+interface CoinRepository {
     fun getMarketList(listener: CoinRemoteDataSourceListener<MarketResponse>)
 
     fun getTickerList(marketNames: String, listener: CoinRemoteDataSourceListener<TickerResponse>)
