+++ wswon/AndroidArchitectureStudy/app/src/main/java/com/architecture/study/data/repository/CoinRepositoryImp.kt
@@ -1,0 +1,28 @@
+package com.architecture.study.data.repository
+
+import com.architecture.study.data.source.CoinRemoteDataSourceImp
+import com.architecture.study.data.source.CoinRemoteDataSourceListener
+import com.architecture.study.network.model.MarketResponse
+import com.architecture.study.network.model.TickerResponse
+
+class CoinRepositoryImp : CoinRepository {
+    companion object {
+        private var instance: CoinRepositoryImp? = null
+        fun getInstance(): CoinRepositoryImp =
+            instance ?: synchronized(this) {
+                instance
+                    ?: CoinRepositoryImp().also {
+                    instance = it
+                }
+            }
+    }
+
+    override fun getMarketList(listener: CoinRemoteDataSourceListener<MarketResponse>) {
+        CoinRemoteDataSourceImp.getInstance().getMarketList(listener)
+    }
+
+    override fun getTickerList(marketNames: String, listener: CoinRemoteDataSourceListener<TickerResponse>) {
+        CoinRemoteDataSourceImp.getInstance()
+            .getTickerList(marketNames, listener)
+    }
+}
